Rel. Garantias - Detalhamento

+++ nds-client/src/main/java/br/com/abril/nds/controllers/cadastro/CotaGarantiaController.java
@@ -373,7 +373,7 @@ public class CotaGarantiaController {
 		
 		List<String> listaMensagens = new ArrayList<String>();
 		
-		if (notaPromissoria.getValor() == null || notaPromissoria.getValor() <= 0) {
+		if (notaPromissoria.getValor() == null || notaPromissoria.getValor().doubleValue() <= 0) {
 			listaMensagens.add("O preenchimento do campo [Valor R$] é obrigatório");
 		}
